requirements.txt
streamlit
pandas
pandas_profiling 
jupyterlab-git
streamlit-aggrid
XlsxWriter
st-gsheets-connection
python-docx